Fix integration tests to support mybatis-spring v1.3.x
Fix incomplete commit done previously.
